Add files via upload
David, I am sorry this is messy. The whole Class1 should be good to go. The only thing from Form1 that we need is the button click information and that code will need to be added to the add form on the generation button however you and Evan need to do that is fine. editSettings should all be added and can tie to Preston's settings form. If we cannot connect it to the app/db,I left some comments of what should make it work.